backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20     # Required for File / Form uploads in FastAPI

# ─── Data & ML ────────────────────────────────────────────────────────────────
pandas==2.2.3
numpy==2.2.5
scikit-learn==1.6.1
joblib==1.4.2
fairlearn==0.12.0

# ─── AI Report Generation ─────────────────────────────────────────────────────
google-genai==1.12.1
python-dotenv==1.1.0

# ─── Dataset Fetching ─────────────────────────────────────────────────────────
ucimlrepo==0.0.7
